検索条件で評価を指定できるようにした

+++ infrastructure/src/main/java/entrance/infrastructure/database/ImageItemView.java
@@ -14,7 +14,7 @@ public class ImageItemView {
     @Column(name = "REGISTERED_DATETIME")
     public LocalDateTime registeredDateTime;
     public String path;
-    public short rank;
+    public int rank;
 
     @Override
     public String toString() {
